+++ prefs/lib/kv_store.dart
@@ -1,0 +1,9 @@
+export 'src/kv_store.dart'
+    show
+        KvStore,
+        KvStoreExt,
+        KvStoreRead,
+        KvStoreWrite,
+        KvStoreGetStringFunction,
+        KvStoreSetStringFunction,
+        KvStoreRemoveFunction;