+++ prefs/lib/prefs_light.dart
@@ -1,3 +1,4 @@
+export 'kv_store.dart';
 export 'src/prefs_light.dart' show PrefsLight, PrefsLightExt;
 export 'src/prefs_light_async.dart' show PrefsLightAsync;
 export 'src/prefs_light_memory.dart' show PrefsMemory;